requirements.txt
streamlit>=1.32.0
google-generativeai>=0.8.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
Pillow>=10.0.0
